Remove 50urls.csv - had MEMEX HT examples not really suitable for general distribution.
Also, testing now uses the 5K URL file from thh-classifiers.
I'll add it here or to a separate project after cleaning:
about 10% of the pages have expired and are for sale, etc.

On 500 of those cases, accuracy and f1 are still around 50%.
It's clear that the bottlenecks are "classified" (needs work) and
"search_engines" (not even considered by this classifier, yet).
Also, that dataset doesn't have "shopping"?